+++ util/src/main/java/org/grnet/status/util/Utility.java
@@ -6,6 +6,13 @@ import jakarta.inject.Inject;
 import jakarta.ws.rs.BadRequestException;
 import org.eclipse.microprofile.config.inject.ConfigProperty;
 
+import java.util.List;
+import java.util.Map;
+import java.util.stream.IntStream;
+
+import static java.lang.Math.min;
+import static java.util.stream.Collectors.toMap;
+
 @ApplicationScoped
 public class Utility {
 
@@ -51,5 +58,21 @@ public class Utility {
 
     public String getUserSurname() {
         return tokenIntrospection.getJsonObject().getString("family_name", null);
+    }
+
+    /**
+     * This method paginates a list of objects.
+     *
+     * @param list The list to be paginated.
+     * @param pageSize The page size.
+     * @return A map containing the pages of objects.
+     */
+    public <T> Map<Integer, List<T>> partition(List<T> list, int pageSize) {
+
+        return IntStream.iterate(0, i -> i + pageSize)
+                .limit((list.size() + pageSize - 1) / pageSize)
+                .boxed()
+                .collect(toMap(i -> i / pageSize,
+                        i -> list.subList(i, min(i + pageSize, list.size()))));
     }
 }